renaming file to fit class name
